requirements/develop.txt
aiohttp==3.7.4.post0
alembic==1.5.8
asyncpg==0.22.0
asyncpgsa==0.27.1
psycopg2==2.8.6
pyyaml==5.4.1